pyproject.toml
[project]
name = "bakohackathon"
version = "0.1.0"
description = "Bakery voice checklist assistant using Gemini Live API"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.0",
]
